# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"EScrapeStatus": {
				"enum": [
					"cancelled",
					"pending",
					"completed"
				],
				"type": "string"
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "savannah",
		"version": "1.0.0",
		"description": "interview test for savannahtech",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Akintola Felix A."
		}
	},
	"paths": {
		"/api/v1/scrape": {
			"post": {
				"operationId": "Scrape",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"status": {
													"$ref": "#/components/schemas/EScrapeStatus"
												},
												"jobId": {
													"type": "string"
												}
											},
											"required": [
												"status",
												"jobId"
											],
											"type": "object"
										},
										"statusCode": {
											"type": "number",
											"format": "double"
										},
										"status": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"statusCode",
										"status",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Scrape"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"themeWord": {
										"type": "string"
									}
								},
								"required": [
									"themeWord"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/scrape/status/{jobId}": {
			"get": {
				"operationId": "GetStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"status": {},
												"jobId": {}
											},
											"required": [
												"status",
												"jobId"
											],
											"type": "object"
										},
										"statusCode": {
											"type": "number",
											"format": "double"
										},
										"status": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"statusCode",
										"status",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Scrape"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "jobId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/scrape/results/{jobId}": {
			"get": {
				"operationId": "GetBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"statusCode": {
											"type": "number",
											"format": "double"
										},
										"status": {
											"type": "boolean"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"statusCode",
										"status",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Scrape"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "jobId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}